requirements.txt
mcp>=1.0.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
